2017 version for Pipe_Qlp.
This version is all about implementing ltl concept to inquiry on a petri net.
This inquiry have it's own syntax so we build a compiler(javacc) to inquiry.
The process will transform the petri net to reachibilty graph and our algorithims will travers and find
the right answer for the inquiry and shows the path and the graph with correct values.